# perf: remove TrieUpdates::removed_nodes and StorageTrieUpdates::removed_nodes (attempt 2)