--- a/crates/chain-state/Cargo.toml
+++ b/crates/chain-state/Cargo.toml
@@ -52,6 +52,7 @@ reth-testing-utils.workspace = true
 alloy-signer.workspace = true
 alloy-signer-local.workspace = true
 alloy-consensus.workspace = true
+criterion.workspace = true
 rand.workspace = true
 
 [features]
@@ -65,3 +66,7 @@ test-utils = [
     "reth-trie/test-utils",
     "revm/test-utils",
 ]
+
+[[bench]]
+name = "trie_state"
+harness = false