Replaces https://github.com/paradigmxyz/reth/pull/13872. This PR is only *slightly* worse than https://github.com/paradigmxyz/reth/pull/13872, but the code changes are minimal.

<details>
<summary>Checklist</summary>

- [x] E2E Benchmarks
- [x] Add `benches/*`
- [x] Improve PR description

</details>

## Motivation

Under high load (TPS > 50,000), `MemoryOverlayStateProviderRef::trie_state` is taking a considerable amount of time (over 200ms). One factor contributing to this is the `TrieUpdates::extend_ref` function. Optimizing the struct definition of `TrieUpdates` could help improve the performance of the `extend_ref` function.

## Solution

First, this PR changes the struct definitions of `TrieUpdates` and `StorageTrieUpdates`:

```diff
pub struct TrieUpdates {
-    pub account_nodes: HashMap<Nibbles, BranchNodeCompact>,
-    pub removed_nodes: HashSet<Nibbles>,
+    pub changed_nodes: HashMap<Nibbles, Option<BranchNodeCompact>>,
     pub storage_tries: B256HashMap<StorageTrieUpdates>,
 }

pub struct StorageTrieUpdates {
     pub is_deleted: bool,
-    pub storage_nodes: HashMap<Nibbles, BranchNodeCompact>,
-    pub removed_nodes: HashSet<Nibbles>,
+    pub changed_nodes: HashMap<Nibbles, Option<BranchNodeCompact>>,
 }
```

Next, this PR replaces the following steps in fn `TrieUpdates::extend_ref`:

1. `self.account_nodes.retain(|nibbles, _| !other.removed_nodes.contains(nibbles));`
2. `self.account_nodes.extend(exclude_empty_from_pair(other.account_nodes.iter().map(|(k, v)| (k.clone(), v.clone()))));`
3. `self.removed_nodes.extend(exclude_empty(other.removed_nodes.iter().cloned()));`

by

1. `self.account_nodes.extend(exclude_empty_from_pair(other.account_nodes.iter().map(|(k, v)| (k.clone(), v.clone()))));`

Similar changes is applied to `StorageTrieUpdates::extend_ref`.

## Criterion Benchmarks

> Note: the current criterion benchmark does not produce very stable results.

<details>
<summary>Outdated</summary>

----
```
[Before]
ERC20                   time:   [503.56 ms 503.85 ms 504.13 ms]
Raw Transfer            time:   [205.55 ms 205.87 ms 206.19 ms]
Uniswap                 time:   [55.124 ms 55.255 ms 55.393 ms]

[After]
ERC20                   time:   [476.01 ms 476.36 ms 476.72 ms]
Raw Transfer            time:   [112.22 ms 112.33 ms 112.49 ms]
Uniswap                 time:   [31.368 ms 31.551 ms 31.738 ms]
```
----

</details>

## E2E Benchmarks

![Total run time for state root calculation](https://github.com/user-attachments/assets/171c70b3-50bd-4490-873c-321f133ddb1f)

(unit: μs) | Before | After | Ratio
-- | -- | -- | --
ERC20 | 559541496 | 515041545 | 0.9204706866
Raw Transfer | 330763242 | 271971019 | 0.8222528518
Uniswap | 1114770107 | 1037325457 | 0.930528591



<details>
<summary>Other benchmarks</summary>

```
"785bc168-9976731c-erc20-30600"
{"tps": 30601.0, "gps": 1056921850.8367347, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-erc20-31000"
{"tps": 31001.0, "gps": 1070736094.1406412, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-erc20-32000"
{"tps": 32001.0, "gps": 1105272921.4311633, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-erc20-33000"
{"tps": 33001.0, "gps": 1139812889.5115511, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-erc20-34000"
{"tps": 34001.0, "gps": 1174346187.1292517, "is_chain_lagged": true, "chain_lag_distance": 4}
"785bc168-9976731c-erc20-36000"
{"tps": 36001.0, "gps": 1243424508.7788463, "is_chain_lagged": true, "chain_lag_distance": 20}
"785bc168-9976731c-erc20-40000"
{"tps": 40001.0, "gps": 1381586516.224, "is_chain_lagged": true, "chain_lag_distance": 61}
"785bc168-9976731c-raw-transfer-54600"
{"tps": 54601.0, "gps": 1146644512.2513661, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-raw-transfer-55000"
{"tps": 55001.0, "gps": 1155044519.4926472, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-raw-transfer-56000"
{"tps": 56001.0, "gps": 1176044515.2560747, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-raw-transfer-57000"
{"tps": 57001.0, "gps": 1197044511.5361216, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-raw-transfer-58000"
{"tps": 58001.0, "gps": 1218044525.3953488, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-raw-transfer-60000"
{"tps": 60001.0, "gps": 1260044514.0, "is_chain_lagged": true, "chain_lag_distance": 5}
"785bc168-9976731c-raw-transfer-64000"
{"tps": 64001.0, "gps": 1344044510.4796574, "is_chain_lagged": true, "chain_lag_distance": 26}
"785bc168-9976731c-uniswap-6330"
{"tps": 6331.0, "gps": 882450529.3786408, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-uniswap-6400"
{"tps": 6401.0, "gps": 892212483.2733675, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-uniswap-6480"
{"tps": 6481.0, "gps": 903381443.1144955, "is_chain_lagged": false, "chain_lag_distance": 1}
"785bc168-9976731c-uniswap-6500"
{"tps": 6501.0, "gps": 906146040.2489707, "is_chain_lagged": false, "chain_lag_distance": 0}
"785bc168-9976731c-uniswap-6540"
{"tps": 6541.0, "gps": 911705913.7027911, "is_chain_lagged": true, "chain_lag_distance": 2}
"785bc168-9976731c-uniswap-6580"
{"tps": 6581.0, "gps": 917291024.3644142, "is_chain_lagged": false, "chain_lag_distance": 1}

"fc28b62f-e11e1f30-bddfae28-erc20-16900/realtime-block.stdout.log"
{"tps": 16901.0, "gps": 583739164.9385916, "is_chain_lagged": false, "chain_lag_distance": 1}
"fc28b62f-e11e1f30-bddfae28-raw-transfer-23000/realtime-block.stdout.log"
{"tps": 23001.0, "gps": 483044502.0552147, "is_chain_lagged": false, "chain_lag_distance": 0}
"fc28b62f-e11e1f30-bddfae28-uniswap-6700/realtime-block.stdout.log"
{"tps": 6701.0, "gps": 934019908.1760107, "is_chain_lagged": true, "chain_lag_distance": 10}

["fc28b62f-a75b99d2-bddfae28-erc20-16900/realtime-block.stdout.log", "515041545 μs"]
{"tps": 16901.0, "gps": 583739134.7166197, "is_chain_lagged": true, "chain_lag_distance": 2}
["fc28b62f-a75b99d2-bddfae28-raw-transfer-23000/realtime-block.stdout.log", "271971019 μs"]
{"tps": 23001.0, "gps": 483044502.34969324, "is_chain_lagged": true, "chain_lag_distance": 7}
["fc28b62f-a75b99d2-bddfae28-uniswap-6700/realtime-block.stdout.log", "1037325457 μs"]
{"tps": 6701.0, "gps": 934018910.3569355, "is_chain_lagged": false, "chain_lag_distance": 1}

["fc28b62f-e11e1f30-412de5a2-erc20-16900/realtime-block.stdout.log", "559541496 μs"]
{"tps": 16901.0, "gps": 583740567.6405634, "is_chain_lagged": true, "chain_lag_distance": 11}
["fc28b62f-e11e1f30-412de5a2-raw-transfer-23000/realtime-block.stdout.log", "330763242 μs"]
{"tps": 23001.0, "gps": 483044502.0828221, "is_chain_lagged": false, "chain_lag_distance": 0}
["fc28b62f-e11e1f30-412de5a2-uniswap-6700/realtime-block.stdout.log", "1114770107 μs"]
{"tps": 6701.0, "gps": 934022094.8027697, "is_chain_lagged": true, "chain_lag_distance": 8}
```
</details>