@kien-rise 2025-02-10
I am gonna reopen this PR. A quick benchmark suggested that it can save 10% of the run time (after https://github.com/paradigmxyz/reth/pull/13976).