@gakonst 2025-03-19
Is this PR still relevant post https://github.com/paradigmxyz/reth/pull/13976?