@kien-rise 2025-03-20
> Is this PR still relevant post #13976?

Sorry for the long period of inactivity. I was focusing on another problem so this PR was left aside.

I believe this PR still have benefits.

My comment on Feb 10 confirmed that there is visible improvement **even after #13976**.